refactor : Seller, ThreePL을 User 컨트롤러로 통합
